# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для расчёта трудовых норм и учёта производственных заказов",
        "title": "Сервис нормирования API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/final/update/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "norm-orders"
                ],
                "summary": "Обновить детали заказа",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заказа для обновления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновлённые данные по заказу",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.UpdateFinalOrderDetails"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешно обновлено"
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка обновления",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/materials/calculation": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recalculate"
                ],
                "summary": "Сохранить новые операции",
                "parameters": [
                    {
                        "description": "Параметры для пересчета норм",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/recalculate_norm.RecalculateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/recalculate_norm.Resp"
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "не удалось пересчитать нормы",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/orders/cancel": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "norm-orders"
                ],
                "summary": "Отменить статус заказа",
                "parameters": [
                    {
                        "description": "ID корневого изделия",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/update.CancelStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Успешно отменено"
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to cancel order",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/orders/nashchelnik/calc": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "norm-orders"
                ],
                "summary": "Сохранить расчёт нащельника",
                "parameters": [
                    {
                        "description": "Данные расчёта нащельника",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/save.SaveNashchelnikCalcRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.GetOrderDetails"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/orders/order-norm/operations": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "norm-orders"
                ],
                "summary": "Сохранить новые операции",
                "parameters": [
                    {
                        "description": "Нормированные операции к наряду",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.OrderNormDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/save.Response"
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "не удалось сохранить нормировку / не удалось сохранить операции",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/orders/order/norm/update/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "norm-orders"
                ],
                "summary": "Обновить нормы по операциям",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID заказа для обновления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновлённые данные по операциям",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.UpdateOrderDetails"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ID обновлённого заказа",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "400": {
                        "description": "Invalid request body",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/workers/all": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workers"
                ],
                "summary": "Получить всех сотрудников",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Бригада сотрудников",
                        "name": "type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.GetWorkers"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "recalculate_norm.RecalculateRequest": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "order_num": {
                    "type": "string"
                },
                "permis_dop_material": {
                    "type": "boolean"
                },
                "position": {
                    "type": "integer"
                },
                "template": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "recalculate_norm.Resp": {
            "type": "object",
            "properties": {
                "context": {
                    "$ref": "#/definitions/recalculate.Context"
                },
                "operation": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.Operation"
                    }
                }
            }
        },
        "save.Response": {
            "type": "object",
            "properties": {
                "order_id": {
                    "type": "integer"
                }
            }
        },
        "save.SaveNashchelnikCalcRequest": {
            "type": "object",
            "properties": {
                "a": {
                    "type": "number"
                },
                "b": {
                    "type": "number"
                },
                "c": {
                    "type": "number"
                },
                "count": {
                    "type": "number"
                },
                "d": {
                    "type": "number"
                },
                "legacy_id": {
                    "type": "integer"
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.NormOperation"
                    }
                },
                "order_num": {
                    "type": "string"
                },
                "sqr": {
                    "type": "number"
                }
            }
        },
...
        "storage.AssignedWorker": {
            "type": "object",
            "properties": {
                "actual_minutes": {
                    "type": "number"
                },
                "actual_value": {
                    "type": "number"
                },
                "employee_id": {
                    "type": "integer"
                }
            }
        },
...
        "storage.GetOrderDetails": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "customer": {
                    "type": "string"
                },
                "head_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.NormOperation"
                    }
                },
                "order_num": {
                    "type": "string"
                },
                "parent_assembly": {
                    "type": "string"
                },
                "parent_product_id": {
                    "type": "integer"
                },
                "part_type": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "profile": {
                    "type": "string"
                },
                "ready_date": {
                    "type": "string"
                },
                "sqr": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "systema": {
                    "type": "string"
                },
                "template_code": {
                    "type": "string"
                },
                "total_time": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                },
                "type_izd": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "storage.GetWorkers": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "storage.NormOperation": {
            "type": "object",
            "properties": {
                "assign_workers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.AssignedWorker"
                    }
                },
                "count": {
                    "type": "number"
                },
                "minutes": {
                    "type": "number"
                },
                "operation_label": {
                    "type": "string"
                },
                "operation_name": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "storage.Operation": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "number"
                },
                "depends_on": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "formula": {
                    "type": "string"
                },
                "group": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "minutes": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "required": {
                    "type": "boolean"
                },
                "show_calc": {
                    "type": "boolean"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        },
...
        "storage.OrderNormDetails": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "number"
                },
                "customer": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.NormOperation"
                    }
                },
                "order_num": {
                    "type": "string"
                },
                "parent_assembly": {
                    "type": "string"
                },
                "parent_product_id": {
                    "type": "integer"
                },
                "part_type": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                },
                "profile": {
                    "type": "string"
                },
                "sqr": {
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "systema": {
                    "type": "string"
                },
                "template_code": {
                    "type": "string"
                },
                "total_time": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                },
                "type_izd": {
                    "type": "string"
                }
            }
        },
...
        "storage.UpdateFinalOrderDetails": {
            "type": "object",
            "properties": {
                "brigade": {
                    "type": "string"
                },
                "coefficient": {
                    "type": "number"
                },
                "customer_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "norm_money": {
                    "type": "number"
                },
                "parent_assembly": {
                    "type": "string"
                },
                "profile": {
                    "type": "string"
                },
                "sqr": {
                    "type": "number"
                },
                "sqr_stv": {
                    "type": "number"
                },
                "systema": {
                    "type": "string"
                },
                "type_izd": {
                    "type": "string"
                }
            }
        },
        "storage.UpdateOrderDetails": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.NormOperation"
                    }
                },
                "order_num": {
                    "type": "string"
                },
                "parent_assembly": {
                    "type": "string"
                },
                "parent_product_id": {
                    "type": "integer"
                },
                "part_type": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_time": {
                    "type": "number"
                },
                "type": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "update.CancelStatusRequest": {
            "type": "object",
            "properties": {
                "root_product_id": {
                    "type": "integer"
                }
            }
        },
...
    }
}