# http-server/recalculate-norm/calculate.go
package recalculate_norm

import (
	"context"
	"encoding/json"
	"log/slog"
	"net/http"
	"time"
	"vue-golang/internal/service/recalculate"
	"vue-golang/internal/storage"

	"github.com/go-chi/render"
)

type NormCalculator interface {
	CalculateNorm(ctx context.Context, orderNum string, pos int, typeIzd string, templateCode string, itemCount int, permisDopMaterial bool) ([]storage.Operation, recalculate.Context, error)
}

type Resp struct {
	Operation []storage.Operation `json:"operation"`
	Context   recalculate.Context `json:"context"`
}

type RecalculateRequest struct {
	OrderNum          string `json:"order_num"`
	Position          int    `json:"position"`
	TypeIzd           string `json:"type"`
	TemplateCode      string `json:"template"`
	ItemCount         int    `json:"count"`
	PermisDopMaterial bool   `json:"permis_dop_material"`
}

// CalculateNormOperations пересчет и сохранение новых норм
// @Summary Сохранить новые операции
// @Tags recalculate
// @Accept json
// @Produce json
// @Param request body RecalculateRequest true "Параметры для пересчета норм"
// @Success 200 {object} Resp
// @Failure 400 {string} string "Неверные данные"
// @Failure 500 {string} string "не удалось пересчитать нормы"
// @Router /api/materials/calculation [post]
func CalculateNormOperations(log *slog.Logger, calc NormCalculator) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		const op = "handler.norm.CalculateNormOperations"

		var req RecalculateRequest

		if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
			http.Error(w, "Некорректный JSON", http.StatusBadRequest)
			return
		}

		if req.TypeIzd == "door" {
			req.PermisDopMaterial = true
		}

		log.Info("Calculating norm", "type", req.TypeIzd, "order", req.OrderNum, "pos", req.Position, "permis", req.PermisDopMaterial)

		ctx, cancel := context.WithTimeout(r.Context(), 10*time.Second)
		defer cancel()

		norm, ctxData, err := calc.CalculateNorm(ctx, req.OrderNum, req.Position, req.TypeIzd, req.TemplateCode, req.ItemCount, req.PermisDopMaterial)
		if err != nil {
			log.With(slog.String("op", op), slog.String("error", err.Error())).Error("ошибка автоматического пересчета норм")
			http.Error(w, "Internal error", http.StatusInternalServerError)
			return
		}

		render.JSON(w, r, Resp{
			Operation: norm,
			Context:   ctxData,
		})
	}
}